requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
python-dotenv==1.0.1
pytz==2024.1
Pillow==10.3.0
